remove RSS references; rename components to not use RSS; RSS is inaccurate; we are not consuming an RSS feed
