+++ config/deploy-azure-storage.json
@@ -2,6 +2,6 @@
   "$schema": "https://developer.microsoft.com/json-schemas/spfx-build/deploy-azure-storage.schema.json",
   "workingDir": "./release/assets/",
   "account": "<!-- STORAGE ACCOUNT NAME -->",
-  "container": "sharepoint-wordpress-rss-feed",
+  "container": "sharepoint-wordpress-news-feed",
   "accessKey": "<!-- ACCESS KEY -->"
 }
