+++ config/package-solution.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://developer.microsoft.com/json-schemas/spfx-build/package-solution.schema.json",
   "solution": {
-    "name": "SPFx WordPress RSS Feed",
+    "name": "SPFx WordPress News Feed",
     "id": "cb04c1be-2cd9-44aa-aadb-2565f1760c25",
     "version": "2.1.0",
     "includeClientSideAssets": true,
@@ -16,10 +16,10 @@
     },
     "metadata": {
       "shortDescription": {
-        "default": "sharepoint-wordpress-rss-feed description"
+        "default": "display a WordPress news feed"
       },
       "longDescription": {
-        "default": "sharepoint-wordpress-rss-feed description"
+        "default": "Provide a target WordPress site URL and display a news feed using its REST API"
       },
       "screenshotPaths": [],
       "videoUrl": "",
@@ -27,14 +27,14 @@
     },
     "features": [
       {
-        "title": "sharepoint-wordpress-rss-feed Feature",
-        "description": "The feature that activates elements of the sharepoint-wordpress-rss-feed solution.",
+        "title": "sharepoint-wordpress-news-feed Feature",
+        "description": "The feature that activates elements of the sharepoint-wordpress-news-feed solution.",
         "id": "83912a2b-270c-4132-aba9-afc07a305ff3",
         "version": "1.0.0.0"
       }
     ]
   },
   "paths": {
-    "zippedPackage": "solution/sharepoint-wordpress-rss-feed.sppkg"
+    "zippedPackage": "solution/sharepoint-wordpress-news-feed.sppkg"
   }
 }
