+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "sharepoint-wordpress-rss-feed",
+  "name": "sharepoint-wordpress-news-feed",
   "version": "0.0.1",
   "private": true,
   "engines": {
