+++ src/webparts/wordPressFeed/assets/welcome-dark.png
