+++ src/webparts/wordPressFeed/assets/welcome-light.png
