+++ src/webparts/wordPressFeed/colorPalette.ts
