+++ src/webparts/wordPressFeed/components/WordPressFeed.module.scss
@@ -1,6 +1,6 @@
 @import '~@fluentui/react/dist/sass/References.scss';
 
-.wordPressRssFeed {
+.wordPressFeed {
   overflow: hidden;
   padding: 1em;
   color: "[theme:bodyText, default: #323130]";
