+++ src/webparts/wordPressFeed/controls/NumericTextField/INumericTextField.ts
