+++ src/webparts/wordPressFeed/controls/ThemePaletteColorPicker/IThemePaletteColorPicker.ts
