+++ src/webparts/wordPressFeed/controls/ThemePaletteColorPicker/PropertyPaneNumericTextField.ts
