+++ src/webparts/wordPressFeed/defaults.ts
