+++ src/webparts/wordPressFeed/loc/en-us.js
